chore: LottoMachine 클래스 파일 분리
