+++ src/App.js
@@ -1,29 +1,12 @@
-const MissionUtils = require("@woowacourse/mission-utils");
-
-class LottoMachine {
-  #lottoList;
-
-  inputPurchaseMoney() {
-    MissionUtils.Console.readLine("구입금액을 입력해 주세요.\n",(inputMoney)=>{
-    })
-  }
-
-  makeLotto() {
-    const lotto = new Set();
-    while(lotto.size < 6){
-      lotto.add(MissionUtils.Random.pickNumberInRange(1, 45));
-    }
-    return [...lotto]
-  }
-}
+const LottoMachine = require("./LottoMachine");
 
 class App {
   constructor() {
-    this.lottomanage = new LottoManage();
+    this.lottomachine = new LottoMachine();
   }
 
   play() {
-    this.lottomanage.inputPurchaseMoney();
+    this.lottomachine.inputPurchaseMoney();
   }
 }
